third_party/WebKit/LayoutTests/http/tests/security/suborigins/crossorigin/suborigin-cross-origin-worker-onerror-no-cors-expected.txt

Ensure that scripts imported into a Worker from cross-origin hosts trigger sanitized onerror messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS worker = new Worker("/workers/resources/worker-importscripts-onerror-sameorigin.js") threw exception SecurityError: Failed to construct 'Worker': Script at 'http://127.0.0.1:8000/workers/resources/worker-importscripts-onerror-sameorigin.js' cannot be accessed from origin 'http://foobar_127.0.0.1:8000'..
PASS successfullyParsed is true

TEST COMPLETE
